Add support for transferring compressed blobs via ByteStream
In many cases, it is desirable for blobs to be sent in compressed form
to and from the cache. While gRPC supports channel-level compression,
the generated bindings APIs require that implementers provide data in
unserialized and uncompressed form.

By allowing compressed data at the REAPI level instead, we can avoid
re-compressing the same data on each request. The ByteStream API stands
to benefit the most from this, with the least amount of effort.

Thanks to Eric Burnett and Grzegorz Lukasik for helping with this.

Implements #147.